webapp/requirements.txt
asgiref==3.7.0
Django==4.2.1
django-model-utils==4.3.1
django-rest-swagger==2.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.20.0
pytz==2023.3
sqlparse==0.4.4
